[REFACTOR] Move common pages into upper folder
